refactor(explorer): rename ContextMenu file
