+++ packages/dm-core-plugins/src/explorer/components/dialogs/NewEntityDialog.tsx
@@ -11,7 +11,7 @@ import { toast } from 'react-toastify'
 import {
   STANDARD_DIALOG_HEIGHT,
   STANDARD_DIALOG_WIDTH,
-} from '../context-menu/ContextMenu'
+} from '../context-menu/NodeRightClickMenu'
 
 type TProps = {
   setDialogId: (id: string) => void
